The Main Compiles
I removed basWorker1.eat(), sick(), and reboot() from the main becuase those methods are not implemented in the BaseWorker class. In the TempWorker class I changed the constructor from SuperWorker(){} to TempWorker(){}. In the SuperWorker class I moved a portion of the print statement down to the next line so it wouldn't pass the page limit. I fixed manager because there was a bunch of random numbers and symbols in it for some reason. I suggest that somebody checks the Manager class because I may have missed some things in it.